refactor: Extract debug panel from PrimeDashboard to separate component
